some tweaks for demo
